requirements.txt
opencv-python==4.9.0.80
fastai==1.0.61
scikit-image==0.22.0
streamlit==1.31.1
streamlit-drawable-canvas==0.9.3
manga-ocr==0.1.11
deepl==1.17.0
python-dotenv==1.0.1
easydict==1.9.0
scikit-learn==1.4.1.post1
pytorch_lightning==2.2.0.post0
kornia==0.5.0
hydra-core==1.3.2
albumentations==0.5.2
webdataset==0.2.86
fire==0.5.0
fugashi==1.3.0
jaconv==0.3.4
loguru==0.7.2
Pillow==10.2.0
pyperclip==1.8.2
transformers>=4.25.0
unidic_lite==1.0.8
ruamel.yaml==0.18.6